Add files to process sentences from Marco's script
In this commit are two files, the file annotated_data contains a class
called Data that can be used to load a txt file with sentences generated
by Marco's script. The function to do so is add_corpus, which will go
through the lines in the file and generate a list of dictionaries with
keys corresponding to the different columns in the data. Additionaly,
the length of the sentence is added as a key. This function can also
be used with different columns and different data files.
The activation data should then be added through a different function
(add_activation_data), of which the helper function should still be
implemented.

There are two additional functions availabe, that allow to write the
data to a pickle file, and to filter the data, according to one of the
values in the dictionary.

Furthermore, there are the signatures of functions to split the data, as
well as get it in the right format for the regression. The functions
should still be implemented.

The script organise_marco_sentences.py provides a simple example of how
to use the Data class on a simple test file (not provided).