Add option to load data from file
The script now has an extra variable called `data_file`.
This file has the same name as the txt_file containing the sentences,
but also contains the *n* used to generate it in the filename.
Before generating the Data object, it is checked if this file exists.
If so, the Data object is loaded from this file, rather than being
recomputed. If not, the Data object is stored to this filename for
future use.