Add functionality to decorrelate data
Add a function to Data class that allows to filter according to a
given number of (pos, depth) pairs within a rectangle specified
by 4 input coordinates.
The function filters all data in the sentence dictionary accordingly.

The function to determine the optimal rectangle coordinates as well as
the maximum number of datapoints to be fetched from there should still
be written.

Should be checked by someone else!